refactor fetching functions
