+++ databases/catdat/scripts/config.ts
@@ -5,11 +5,9 @@ export const PLURALS = {
 	functor: 'functors',
 } as const
 
-// TODO: move this to the schema
 export const STRUCTURE_MAPS = [
 	['source', 'functor', 'category'],
 	['target', 'functor', 'category'],
 ] as const
 
-// TODO: move this to the schema
 export const STRUCTURES_WITH_DUALS: StructureType[] = ['category']
