+++ src/lib/server/fetchers/properties.ts
@@ -1,6 +1,6 @@
 import type { GroupedPropertyShort, StructureType } from '$lib/commons/types'
 import sql from 'sql-template-tag'
-import { query } from '../db.catdat'
+import { query } from '$lib/server/db.catdat'
 import { error } from '@sveltejs/kit'
 
 export function get_property_ids(type: StructureType) {
@@ -19,12 +19,13 @@ export function get_property_ids(type: StructureType) {
 
 export function get_grouped_properties(type: StructureType) {
 	const { rows: properties, err } = query<GroupedPropertyShort>(sql`
-		SELECT id, relation, dual_property_id FROM properties
+		SELECT id, relation, dual_property_id
+		FROM properties
 		WHERE type = ${type}
 		ORDER BY lower(id)
 	`)
 
-	if (err) error(500, 'Could not load properties')
+	if (err) error(500, 'Failed to load properties')
 
 	const seen = new Set()
